Added moment library
