+++ package.json
@@ -20,6 +20,7 @@
     "@supabase/supabase-js": "^2.97.0",
     "@tanstack/react-query": "^5.90.21",
     "@upstash/redis": "^1.36.3",
+    "moment": "^2.30.1",
     "next": "16.1.6",
     "react": "19.2.3",
     "react-dom": "19.2.3"
