+++ src/app/dashboard/[campaignId]/DashboardLayout.tsx
@@ -5,8 +5,22 @@ import { useParams, useRouter, usePathname, notFound } from "next/navigation";
 import Navbar from "@/src/components/Navbar";
 import DashboardTabs from "@/src/components/dashboard/DashboardTabs";
 import { getCampaignById, sampleCampaigns } from "../sampleCampaigns";
+import { TotalRaisedCard } from "@/src/components/dashboard/TotalRaisedCard";
+import { TotalDonorsCard } from "@/src/components/dashboard/TotalDonorsCard";
+import { DaysRemainingCard } from "@/src/components/dashboard/DaysRemainingCard";
+import { mockAnalyticsData } from "@/src/app/dashboard/mockAnalyticsData";
+import moment from "moment";
 
 type DashboardTab = "Overview" | "Donors" | "Analytics";
+
+function getDaysRemaining(endDateIso: string) {
+  const today = moment().startOf("day");
+  const end = moment(endDateIso, "YYYY-MM-DD");
+
+  const diff = end.diff(today, "days");
+
+  return Math.max(0, diff);
+}
 
 export default function DashboardShell({ children }: { children: ReactNode }) {
   const router = useRouter();
@@ -46,6 +60,23 @@ export default function DashboardShell({ children }: { children: ReactNode }) {
       <div className="flex-1 bg-gray-50 p-10">
         <h3 className="text-4xl font-bold text-[#096B2E]">{campaign.name}</h3>
         <DashboardTabs selectedTab={selectedTab} onChange={handleTabChange} />
+        <div className="grid grid-cols-1 md:grid-cols-2 gap-6">
+          <TotalRaisedCard
+            totalRaised={mockAnalyticsData.totalRaised}
+            campaignGoal={mockAnalyticsData.campaignGoal}
+            raisedChangePercent={mockAnalyticsData.raisedChangePercent}
+          />
+          <div className="flex flex-col gap-6">
+            <TotalDonorsCard
+              totalDonors={mockAnalyticsData.totalDonors}
+              donorsChangePercent={mockAnalyticsData.donorsChangePercent}
+            />
+            <DaysRemainingCard
+              daysRemaining={getDaysRemaining(mockAnalyticsData.campaignEndDate)}
+            />
+          </div>
+        </div>
+        
         {children}
       </div>
     </div>
